requirements.txt
flask == 2.2.2
flask-sqlalchemy == 3.0.2
psycopg2-binary == 2.9.6
python-dotenv == 0.21.0
pytest == 7.2.1